@@ -6,18 +6,19 @@
   "private": true,
   "type": "module",
   "dependencies": {
-    "@tuleap/tlp-fetch": "workspace:*",
-    "@tuleap/vue-permissions-per-group-badge": "workspace:*",
-    "@tuleap/vue2-gettext-init": "workspace:*",
-    "vue": "^2.7.8",
-    "vue-gettext": "^2.1.12"
+    "@tuleap/fault": "workspace:*",
+    "@tuleap/fetch-result": "workspace:*",
+    "@tuleap/vue3-permissions-per-group-badge": "workspace:*",
+    "@tuleap/vue3-gettext-init": "workspace:*",
+    "vue": "3.3.4",
+    "vue3-gettext": "^2.4.0"
   },
   "devDependencies": {
     "@tuleap/build-system-configurator": "workspace:*",
     "@tuleap/po-gettext-plugin": "workspace:*",
-    "@vue/test-utils": "^1.3",
-    "vue-class-component": "^7.2.6",
-    "vue-property-decorator": "^9.1.2"
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vue/test-utils": "^2.2.10",
+    "neverthrow": "6.0.0"
   },
   "scripts": {
     "typecheck": "vue-tsc --noEmit",
