@@ -52,10 +52,12 @@ class PermissionPerGroupController extends RouterLink
         if (! $request->getCurrentUser()->isAdmin($request->getProject()->getID())) {
             $GLOBALS['Response']->send400JSONErrors(
                 [
-                    'error' => dgettext(
-                        'tuleap-git',
-                        "You don't have permissions to see user groups."
-                    ),
+                    'error' => [
+                        'message' => dgettext(
+                            'tuleap-git',
+                            "You don't have permissions to see user groups."
+                        ),
+                    ],
                 ]
             );
         }
