@@ -17,28 +17,23 @@
  * along with Tuleap. If not, see http://www.gnu.org/licenses/.
  */
 
-import type { Wrapper } from "@vue/test-utils";
 import { shallowMount } from "@vue/test-utils";
 import GitPermissionsTable from "./GitPermissionsTable.vue";
-import localVueForTest from "./helper/local-vue-for-test";
+import { createGettext } from "vue3-gettext";
+import type { RepositoryFineGrainedPermissions } from "./type";
 
 describe("GitPermissionsTable", () => {
-    let propsData = {};
-
-    function instantiateComponent(): Wrapper<GitPermissionsTable> {
-        return shallowMount(GitPermissionsTable, {
-            propsData,
-            localVue: localVueForTest,
-        });
-    }
-
     it("When there are no repositories, Then empty state is displayed", () => {
-        propsData = {
-            repositories: [],
-            selectedUgroupName: "Project Member",
-        };
-
-        const wrapper = instantiateComponent();
+        const wrapper = shallowMount(GitPermissionsTable, {
+            props: {
+                repositories: [],
+                selected_ugroup_name: "Project Member",
+                filter: "",
+            },
+            global: {
+                plugins: [createGettext({ silent: true })],
+            },
+        });
 
         expect(wrapper.find("[data-test=git-permission-table-empty-state]").text()).toBe(
             "Project Member has no permission for any repository in this project",
@@ -46,12 +41,16 @@ describe("GitPermissionsTable", () => {
     });
 
     it("When there are no repository and no ugroup selected, Then empty state is displayed", () => {
-        propsData = {
-            repositories: [],
-            selectedUgroupName: "",
-        };
-
-        const wrapper = instantiateComponent();
+        const wrapper = shallowMount(GitPermissionsTable, {
+            props: {
+                repositories: [],
+                selected_ugroup_name: "",
+                filter: "",
+            },
+            global: {
+                plugins: [createGettext({ silent: true })],
+            },
+        });
 
         expect(wrapper.find("[data-test=git-permission-table-empty-state]").text()).toBe(
             "No repository found for project",
@@ -59,15 +58,21 @@ describe("GitPermissionsTable", () => {
     });
 
     it("When all repository are hidden, Then empty state is displayed", async () => {
-        propsData = {
-            repositories: [{ name: "repo" }],
-            selectedUgroupName: "",
-        };
-
-        const wrapper = instantiateComponent();
-        wrapper.setData({
-            nb_repo_hidden: 1,
+        const wrapper = shallowMount(GitPermissionsTable, {
+            props: {
+                repositories: [
+                    { repository_id: 1, name: "repo" } as RepositoryFineGrainedPermissions,
+                ],
+                selected_ugroup_name: "",
+                filter: "",
+            },
+            global: {
+                plugins: [createGettext({ silent: true })],
+            },
         });
+        wrapper
+            .find("[data-test=git-permissions-table-repository-1]")
+            .trigger("filtered", { hidden: true });
 
         await wrapper.vm.$nextTick();
 
